pyproject.toml
[project]
name = "resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.2.1",
    "google-generativeai>=0.8.5",
    "psycopg2-binary>=2.9.11",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
]
